+++ rich/progress.py
@@ -27,7 +27,7 @@ from typing import (
 )
 
 from . import filesize, get_console
-from .bar import Bar
+from .progress_bar import ProgressBar
 from .console import (
     Console,
     ConsoleRenderable,
@@ -47,7 +47,6 @@ from .text import Text
 TaskID = NewType("TaskID", int)
 
 ProgressType = TypeVar("ProgressType")
-
 
 GetTimeCallable = Callable[[], float]
 
@@ -146,9 +145,10 @@ def track(
         refresh_per_second=refresh_per_second,
     )
 
-    yield from progress.track(
-        sequence, total=total, description=description, update_period=update_period
-    )
+    with progress:
+        yield from progress.track(
+            sequence, total=total, description=description, update_period=update_period
+        )
 
 
 class ProgressColumn(ABC):
@@ -243,9 +243,9 @@ class BarColumn(ProgressColumn):
         self.pulse_style = pulse_style
         super().__init__()
 
-    def render(self, task: "Task") -> Bar:
+    def render(self, task: "Task") -> ProgressBar:
         """Gets a progress bar widget for a task."""
-        return Bar(
+        return ProgressBar(
             total=max(0, task.total),
             completed=max(0, task.completed),
             width=None if self.bar_width is None else max(1, self.bar_width),
@@ -292,15 +292,30 @@ class TotalFileSizeColumn(ProgressColumn):
 
 
 class DownloadColumn(ProgressColumn):
-    """Renders file size downloaded and total, e.g. '0.5/2.3 GB'."""
+    """Renders file size downloaded and total, e.g. '0.5/2.3 GB'.
+
+    Args:
+        binary_units (bool, optional): Use binary units, KiB, MiB etc. Defaults to False.
+    """
+
+    def __init__(self, binary_units: bool = False) -> None:
+        self.binary_units = binary_units
+        super().__init__()
 
     def render(self, task: "Task") -> Text:
         """Calculate common unit for completed and total."""
         completed = int(task.completed)
         total = int(task.total)
-        unit, suffix = filesize.pick_unit_and_suffix(
-            total, ["bytes", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"], 1000
-        )
+        if self.binary_units:
+            unit, suffix = filesize.pick_unit_and_suffix(
+                total,
+                ["bytes", "KiB", "MiB", "GiB", "TiB", "PiB", "EiB", "ZiB", "YiB"],
+                1024,
+            )
+        else:
+            unit, suffix = filesize.pick_unit_and_suffix(
+                total, ["bytes", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"], 1000
+            )
         completed_ratio = completed / unit
         total_ratio = total / unit
         precision = 0 if unit == 1 else 1
@@ -436,6 +451,10 @@ class Task:
         estimate = ceil(self.remaining / speed)
         return estimate
 
+    def _reset(self) -> None:
+        """Reset progress."""
+        self._progress.clear()
+
 
 class _RefreshThread(Thread):
     """A thread that calls refresh() on the Process object at regular intervals."""
@@ -499,8 +518,8 @@ class Progress(JupyterMixin, RenderHook):
         refresh_per_second (Optional[int], optional): Number of times per second to refresh the progress information or None to use default (10). Defaults to None.
         speed_estimate_period: (float, optional): Period (in seconds) used to calculate the speed estimate. Defaults to 30.
         transient: (bool, optional): Clear the progress on exit. Defaults to False.
-        redirect_stout: (bool, optional): Enable redirection of stdout, so ``print`` may be used. Defaults to True.
-        redirect_stout: (bool, optional): Enable redirection of stderr. Defaults to True.
+        redirect_stdout: (bool, optional): Enable redirection of stdout, so ``print`` may be used. Defaults to True.
+        redirect_stderr: (bool, optional): Enable redirection of stderr. Defaults to True.
         get_time: (Callable, optional): A callable that gets the current time, or None to use time.monotonic. Defaults to None.
     """
 
@@ -664,19 +683,19 @@ class Progress(JupyterMixin, RenderHook):
             task_id = self.add_task(description, total=task_total)
         else:
             self.update(task_id, total=task_total)
-        with self:
-            if self.auto_refresh:
-                with _TrackThread(self, task_id, update_period) as track_thread:
-                    for value in sequence:
-                        yield value
-                        track_thread.completed += 1
-            else:
-                advance = self.advance
-                refresh = self.refresh
+
+        if self.auto_refresh:
+            with _TrackThread(self, task_id, update_period) as track_thread:
                 for value in sequence:
                     yield value
-                    advance(task_id, 1)
-                    refresh()
+                    track_thread.completed += 1
+        else:
+            advance = self.advance
+            refresh = self.refresh
+            for value in sequence:
+                yield value
+                advance(task_id, 1)
+                refresh()
 
     def start_task(self, task_id: TaskID) -> None:
         """Start a task.
@@ -737,6 +756,7 @@ class Progress(JupyterMixin, RenderHook):
 
             if total is not None:
                 task.total = total
+                task._reset()
             if advance is not None:
                 task.completed += advance
             if completed is not None:
@@ -760,7 +780,44 @@ class Progress(JupyterMixin, RenderHook):
                 popleft()
             while len(_progress) > 1000:
                 popleft()
-            _progress.append(ProgressSample(current_time, update_completed))
+            if update_completed > 0:
+                _progress.append(ProgressSample(current_time, update_completed))
+
+    def reset(
+        self,
+        task_id: TaskID,
+        *,
+        start: bool = True,
+        total: Optional[int] = None,
+        completed: int = 0,
+        visible: Optional[bool] = None,
+        description: Optional[str] = None,
+        **fields: Any,
+    ) -> None:
+        """Reset a task so completed is 0 and the clock is reset.
+
+        Args:
+            task_id (TaskID): ID of task.
+            start (bool, optional): Start the task after reset. Defaults to True.
+            total (int, optional): New total steps in task, or None to use current total. Defaults to None.
+            completed (int, optional): Number of steps completed. Defaults to 0.
+            **fields (str): Additional data fields required for rendering.
+        """
+        current_time = self.get_time()
+        with self._lock:
+            task = self._tasks[task_id]
+            task._reset()
+            task.start_time = current_time if start else None
+            if total is not None:
+                task.total = total
+            task.completed = completed
+            if visible is not None:
+                task.visible = visible
+            if fields:
+                task.fields = fields
+            if description is not None:
+                task.description = description
+            self.refresh()
 
     def advance(self, task_id: TaskID, advance: float = 1) -> None:
         """Advance task by a number of steps.