+++ rich/table.py
@@ -48,9 +48,16 @@ class Column:
     """str: How to justify text within the column ("left", "center", "right", or "full")"""
 
     overflow: "OverflowMethod" = "ellipsis"
+    """str: Overflow method."""
 
     width: Optional[int] = None
     """Optional[int]: Width of the column, or ``None`` (default) to auto calculate width."""
+
+    min_width: Optional[int] = None
+    """Optional[int]: Minimum width of column, or ``None`` for no minimum. Defaults to None."""
+
+    max_width: Optional[int] = None
+    """Optional[int]: Maximum width of column, or ``None`` for no maximum. Defaults to None."""
 
     ratio: Optional[int] = None
     """Optional[int]: Ratio to use when calculating column width, or ``None`` (default) to adapt to column contents."""
@@ -121,6 +128,7 @@ class Table(JupyterMixin):
         title: TextType = None,
         caption: TextType = None,
         width: int = None,
+        min_width: int = None,
         box: Optional[box.Box] = box.HEAVY_HEAD,
         safe_box: Optional[bool] = None,
         padding: PaddingDimensions = (0, 1),
@@ -155,6 +163,7 @@ class Table(JupyterMixin):
         self.title = title
         self.caption = caption
         self.width = width
+        self.min_width = min_width
         self.box = box
         self.safe_box = safe_box
         self._padding = Padding.unpack(padding)
@@ -246,13 +255,12 @@ class Table(JupyterMixin):
 
         extra_width = self._extra_width
 
-        max_width = sum(self._calculate_column_widths(console, max_width))
+        max_width = sum(self._calculate_column_widths(console, max_width - extra_width))
 
         _measure_column = self._measure_column
 
         measurements = [
-            _measure_column(console, column, max_width - extra_width)
-            for column in self.columns
+            _measure_column(console, column, max_width) for column in self.columns
         ]
         minimum_width = (
             sum(measurement.minimum for measurement in measurements) + extra_width
@@ -262,7 +270,9 @@ class Table(JupyterMixin):
             if (self.width is None)
             else self.width
         )
-        return Measurement(minimum_width, maximum_width)
+        measurement = Measurement(minimum_width, maximum_width)
+        measurement = measurement.clamp(self.min_width)
+        return measurement
 
     @property
     def padding(self) -> Tuple[int, int, int, int]:
@@ -286,6 +296,8 @@ class Table(JupyterMixin):
         justify: "JustifyMethod" = "left",
         overflow: "OverflowMethod" = "ellipsis",
         width: int = None,
+        min_width: int = None,
+        max_width: int = None,
         ratio: int = None,
         no_wrap: bool = False,
     ) -> None:
@@ -300,7 +312,9 @@ class Table(JupyterMixin):
             footer_style (Union[str, Style], optional): Style for the header. Defaults to "none".
             style (Union[str, Style], optional): Style for the column cells. Defaults to "none".
             justify (JustifyMethod, optional): Alignment for cells. Defaults to "left".
-            width (int, optional): A minimum width in characters. Defaults to None.
+            width (int, optional): Desired width of column in characters, or None to fit to contents. Defaults to None.
+            min_width (Optional[int], optional): Minimum width of column, or ``None`` for no minimum. Defaults to None.
+            max_width (Optional[int], optional): Maximum width of column, or ``None`` for no maximum. Defaults to None.
             ratio (int, optional): Flexible ratio for the column (requires ``Table.expand`` or ``Table.width``). Defaults to None.
             no_wrap (bool, optional): Set to ``True`` to disable wrapping of this column.
         """
@@ -319,6 +333,8 @@ class Table(JupyterMixin):
             justify=justify,
             overflow=overflow,
             width=width,
+            min_width=min_width,
+            max_width=max_width,
             ratio=ratio,
             no_wrap=no_wrap,
         )
@@ -376,12 +392,10 @@ class Table(JupyterMixin):
         max_width = options.max_width
         if self.width is not None:
             max_width = self.width
-        if self.box:
-            max_width -= len(self.columns) - 1
-            if self.show_edge:
-                max_width -= 2
-        widths = self._calculate_column_widths(console, max_width)
-        table_width = sum(widths) + self._extra_width
+
+        extra_width = self._extra_width
+        widths = self._calculate_column_widths(console, max_width - extra_width)
+        table_width = sum(widths) + extra_width
 
         render_options = options.update(width=table_width)
 
@@ -417,6 +431,7 @@ class Table(JupyterMixin):
         ]
         widths = [_range.maximum or 1 for _range in width_ranges]
         get_padding_width = self._get_padding_width
+        extra_width = self._extra_width
 
         if self.expand:
             ratios = [col.ratio or 0 for col in columns if col.flexible]
@@ -440,7 +455,9 @@ class Table(JupyterMixin):
 
         if table_width > max_width:
             widths = self._collapse_widths(
-                widths, [not column.no_wrap for column in columns], max_width
+                widths,
+                [(column.width is None and not column.no_wrap) for column in columns],
+                max_width,
             )
             table_width = sum(widths)
 
@@ -456,8 +473,15 @@ class Table(JupyterMixin):
             ]
             widths = [_range.maximum or 1 for _range in width_ranges]
 
-        if table_width < max_width and self.expand:
-            pad_widths = ratio_distribute(max_width - table_width, widths)
+        if (table_width < max_width and self.expand) or (
+            self.min_width is not None and table_width < (self.min_width - extra_width)
+        ):
+            _max_width = (
+                max_width
+                if self.min_width is None
+                else min(self.min_width - extra_width, max_width)
+            )
+            pad_widths = ratio_distribute(_max_width - table_width, widths)
             widths = [_width + pad for _width, pad in zip(widths, pad_widths)]
 
         return widths
@@ -570,7 +594,7 @@ class Table(JupyterMixin):
             # Fixed width column
             return Measurement(
                 column.width + padding_width, column.width + padding_width
-            )
+            ).with_maximum(max_width)
         # Flexible column, we need to measure contents
         min_widths: List[int] = []
         max_widths: List[int] = []
@@ -582,10 +606,15 @@ class Table(JupyterMixin):
             append_min(_min)
             append_max(_max)
 
-        return Measurement(
+        measurement = Measurement(
             max(min_widths) if min_widths else 1,
             max(max_widths) if max_widths else max_width,
+        ).with_maximum(max_width)
+        measurement = measurement.clamp(
+            None if column.min_width is None else column.min_width + padding_width,
+            None if column.max_width is None else column.max_width + padding_width,
         )
+        return measurement
 
     def _render(
         self, console: "Console", options: "ConsoleOptions", widths: List[int]
@@ -683,13 +712,16 @@ class Table(JupyterMixin):
                         _box.get_row(widths, "foot", edge=show_edge), border_style
                     )
                     yield new_line
-                if first:
-                    left, right, divider = box_segments[0]
-                elif last:
-                    left, right, divider = box_segments[2]
-                else:
-                    left, right, divider = box_segments[1]
+                left, right, _divider = box_segments[0 if first else (2 if last else 1)]
 
+                # If the column divider is whitespace also style it with the row background
+                divider = (
+                    _divider
+                    if _divider.text.strip()
+                    else _Segment(
+                        _divider.text, row_style.background_style + _divider.style
+                    )
+                )
                 for line_no in range(max_height):
                     if show_edge:
                         yield left
@@ -706,7 +738,7 @@ class Table(JupyterMixin):
                         yield from rendered_cell[line_no]
                     yield new_line
             if _box and first and show_header:
-                yield Segment(
+                yield _Segment(
                     _box.get_row(widths, "head", edge=show_edge), border_style
                 )
                 yield new_line
@@ -734,25 +766,62 @@ class Table(JupyterMixin):
 
 
 if __name__ == "__main__":  # pragma: no cover
-    from .console import Console
+    from rich.console import Console
+    from rich.highlighter import ReprHighlighter
+    from rich.table import Table
 
-    c = Console()
     table = Table(
-        show_lines=False,
-        row_styles=["red", "green"],
-        expand=False,
-        show_header=True,
-        show_footer=False,
-        show_edge=True,
+        title="Star Wars Movies",
+        caption="Rich example table",
+        caption_justify="right",
     )
-    table.add_column("foo", no_wrap=True, footer="BAR")
-    table.add_column("bar")
-    table.add_column("baz")
-    table.add_row("Magnet", "foo" * 20, "bar" * 10, "egg" * 15)
 
-    for width in range(170, 1, -1):
-        print(" " * width + "<|")
-        c = Console(width=width)
-        c.print(table)
+    table.add_column("Released", header_style="bright_cyan", style="cyan", no_wrap=True)
+    table.add_column("Title", style="magenta")
+    table.add_column("Box Office", justify="right", style="green")
 
-    c.print("Some more words", width=4, overflow="ellipsis")
+    table.add_row("Dec 20, 2019", "Star Wars: The Rise of Skywalker", "$952,110,690")
+    table.add_row("May 25, 2018", "Solo: A Star Wars Story", "$393,151,347")
+    table.add_row("Dec 15, 2017", "Star Wars Ep. V111: The Last Jedi", "$1,332,539,889")
+    table.add_row("Dec 16, 2016", "Rogue One: A Star Wars Story", "$1,332,439,889")
+
+    def header(text: str) -> None:
+        console.print()
+        console.rule(highlight(text))
+        console.print()
+
+    console = Console()
+    highlight = ReprHighlighter()
+    header("Example Table")
+    console.print(table, justify="center")
+
+    table.expand = True
+    header("expand=True")
+    console.print(table, justify="center")
+
+    table.width = 50
+    header("width=50")
+
+    console.print(table, justify="center")
+
+    table.width = None
+    table.expand = False
+    table.row_styles = ["dim", "none"]
+    header("row_styles=['dim', 'none']")
+
+    console.print(table, justify="center")
+
+    table.width = None
+    table.expand = False
+    table.row_styles = ["dim", "none"]
+    table.leading = 1
+    header("leading=1, row_styles=['dim', 'none']")
+    console.print(table, justify="center")
+
+    table.width = None
+    table.expand = False
+    table.row_styles = ["dim", "none"]
+    table.show_lines = True
+    table.leading = 0
+    header("show_lines=True, row_styles=['dim', 'none']")
+    console.print(table, justify="center")