+++ tools/profile_pretty.py
@@ -11,11 +11,13 @@ with open("cats.json") as fh:
     cats = json.load(fh)
 
 
+console.begin_capture()
 start = time()
 pretty = Pretty(cats)
 console.print(pretty, overflow="ignore", crop=False)
+result = console.end_capture()
 taken = (time() - start) * 1000
-
+print(result)
 
 print(console.file.getvalue())
 print(f"{taken:.1f}")