+++ tools/stress_test_pretty.py
@@ -5,7 +5,15 @@ from rich.pretty import Pretty
 DATA = {
     "foo": [1, 2, 3, (), {}, (1, 2, 3), {4, 5, 6, (7, 8, 9)}, "Hello, World"],
     "bar": [None, (False, True)] * 2,
+    "Dune": {
+        "names": {
+            "Paul Atriedies",
+            "Vladimir Harkonnen",
+            "Thufir Haway",
+            "Duncan Idaho",
+        }
+    },
 }
 console = Console()
 for w in range(130):
-    console.print(Panel(Pretty(DATA), width=w))
+    console.print(Panel(Pretty(DATA, indent_guides=True), width=w))